Move notification proxy to its own namespace
